Squashed 'Externals/Core/' changes from 02ec6f5..392236e
392236e decimal to double.

git-subtree-dir: Externals/Core
git-subtree-split: 392236e9837c03d376c30cb8e6e014d6bed25b96

+++ PS.Mothership.Core/PS.Mothership.Core.Common/Dto/Merchant/AcquiringOfferDto.cs
@@ -25,37 +25,37 @@ namespace PS.Mothership.Core.Common.Dto.Merchant
         public int DebitCardTranxMonthly { get; set; }
 
         [DataMember]
-        public decimal CreditCardRate { get; set; }
+        public double CreditCardRate { get; set; }
 
         [DataMember]
-        public decimal DebitCardRate { get; set; }
+        public double DebitCardRate { get; set; }
 
         [DataMember]
-        public decimal PremiumCardRate { get; set; }
+        public double PremiumCardRate { get; set; }
 
         [DataMember]
-        public decimal CommercialCardRate { get; set; }
+        public double CommercialCardRate { get; set; }
 
         [DataMember]
-        public decimal NTCJoiningFee { get; set; }
+        public double NTCJoiningFee { get; set; }
 
         [DataMember]
-        public decimal MMServiceCharge { get; set; }
+        public double MMServiceCharge { get; set; }
 
         [DataMember]
-        public decimal AuthorisationFee { get; set; }
+        public double AuthorisationFee { get; set; }
 
         [DataMember]
         public string AMEXAccountNumber { get; set; }
 
         [DataMember]
-        public decimal MinimumCardTransactionValue { get; set; }
+        public double MinimumCardTransactionValue { get; set; }
 
         [DataMember]
-        public decimal MaximumCardTransactionValue { get; set; }
+        public double MaximumCardTransactionValue { get; set; }
 
         [DataMember]
-        public decimal AverageCardTransactionValue { get; set; }
+        public double AverageCardTransactionValue { get; set; }
 
         [DataMember]
         public bool AMEXNewAccount { get; set; }
@@ -67,7 +67,7 @@ namespace PS.Mothership.Core.Common.Dto.Merchant
         public GenBusinessLegalTypeEnum BusinessLegalTypeKey { get; set; }
 
         [DataMember]
-        public decimal PCIComplianceFee { get; set; }
+        public double PCIComplianceFee { get; set; }
 
         [DataMember]
         public GenRecStatusEnum RecStatusKey { get; set; }
